refactor: change authentication flow
