+++ src/app/models/api/auth.interface.ts
@@ -2,6 +2,7 @@ export interface NonceResponse {
     status: string
     message: string,
     data: {
+        id: string,
         message: string,
         nonce: number
     }
@@ -11,6 +12,7 @@ export interface ValidateSignatureResponse {
     status: string
     message: string,
     data: {
+        id: string,
         token: string,
         registeredUser: boolean
     }
@@ -20,4 +22,12 @@ export interface RegisterRequest {
     wallet: string,
     name: string,
     email: string
+}
+
+export interface RegisterResponse {
+    status: string
+    message: string,
+    data: {
+        token: string
+    }
 }
